+++ xarray/tests/test_dataarray.py
@@ -1639,6 +1639,16 @@ class TestDataArray:
                 expected.indexes[dim_name], actual.indexes[dim_name]
             )
 
+        # as kwargs
+        array = DataArray(np.random.randn(3), {"x": list("abc")}, "x")
+        expected = DataArray(array.values, {"x": ("y", list("abc"))}, dims="y")
+        actual = array.swap_dims(x="y")
+        assert_identical(expected, actual)
+        for dim_name in set().union(expected.indexes.keys(), actual.indexes.keys()):
+            pd.testing.assert_index_equal(
+                expected.indexes[dim_name], actual.indexes[dim_name]
+            )
+
         # multiindex case
         idx = pd.MultiIndex.from_arrays([list("aab"), list("yzz")], names=["y1", "y2"])
         array = DataArray(np.random.randn(3), {"y": ("x", idx)}, "x")